# backend/docs/API_DOCUMENTATION.yaml
openapi: 3.0.3
info:
  title: CozyVTT Backend Route Reference
  description: |
    The HTTP routes CozyVTT's own web client calls, described for people working
    on CozyVTT itself.

    ## No stability guarantee
    This is an internal surface, not a public API. It is **not versioned**, it
    carries **no compatibility promise**, and any route here may change shape or
    disappear in a point release without notice or a deprecation period.

    There are no API keys, no bearer tokens and no service accounts. The only
    credential is the session cookie described below — but a program can obtain
    one exactly as the web client does, by posting a user's own email and
    password to `/api/auth/login` and keeping the cookie it returns. That cookie
    also authenticates the Socket.io connection, which shares the same session.

    So scripting CozyVTT is possible, and people have done it. What it does not
    come with is a promise: a script authenticates **as a specific user** and can
    do only what that user could do in the browser, the session expires after an
    hour of inactivity (refreshed by activity), an account with MFA enabled needs
    the second step, and any route here may change shape in a point release
    without notice. Treat this document as a map of the current code rather than
    a contract.

    ## Authentication
    All endpoints (except `/health`, `/api/setup/*`, and public auth routes) require authentication via session cookies.
...

...

...

...

...

...

...

...
servers:
...
security:
  - cookieAuth: []
...
paths:
...
  /api/setup/init:
    post:
      tags:
        - Setup
      summary: Initialize system
      description: Create first admin account and complete setup wizard. Only works if setup is not yet complete.
      operationId: initializeSystem
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - displayName
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 12
                displayName:
                  type: string
                instanceName:
                  type: string
                  maxLength: 100
                  default: CozyVTT
                  description: >-
                    Shown in browser tabs and page titles. Blank, missing or
                    non-string values leave the default in place rather than
                    failing setup, which can only be run once.
                timezone:
                  type: string
                  maxLength: 50
                  default: UTC
                  description: IANA timezone name, used for timestamps.
                allowRegistration:
                  type: boolean
                  default: false
                  description: >-
                    Whether people may create their own accounts. Changeable
                    afterwards from Admin → Settings.
      responses:
        '201':
          description: Setup completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ============================================
  # AUTHENTICATION
  # ============================================
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account. First user automatically becomes ADMIN. Rate limited (5/15min).
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - displayName
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 12
                displayName:
                  type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cozyvtt.sid
      description: Session cookie obtained via /api/auth/login
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        platformRole:
          type: string
          enum: [ADMIN, USER]
        mfaEnabled:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: BadRequest
          description: Short machine-readable error identifier
        message:
          type: string
          example: One or more fields failed validation.
          description: Human-readable error description
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: Optional structured details (e.g. per-field Zod issues)
...
  responses:
    BadRequest:
      description: Bad Request - Invalid input or validation failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: BadRequest
            message: Required field "name" is missing.
...
    RateLimited:
      description: |
        Too Many Requests - Rate limit exceeded. Response includes
        `RateLimit-Limit`, `RateLimit-Remaining`, and `RateLimit-Reset` headers.
      headers:
        RateLimit-Limit:
          schema: { type: integer }
          description: Maximum requests permitted in the current window
        RateLimit-Remaining:
          schema: { type: integer }
          description: Requests remaining in the current window
        RateLimit-Reset:
          schema: { type: integer }
          description: Seconds until the window resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: TooManyRequests
            message: Upload rate limit reached. Please slow down and try again in a minute.
...